requirements.txt
django-dotenv
pint